Add files via upload
The Final Code Behavioural_Cloning.py which is being used simply use Polynomial Regression to predict the correct steering and throttle of the car whose training data was recorded using the UDACITY simulator.
Link for the Simulator-  https://github.com/udacity/self-driving-car-sim

Traffic_Sign_Classification is not being used in the final code but can be used in future model with more parameters like what should our self driving car do when it detect those signs.

Finally to test your model on the autonomous mode in the simulator, firstly you need to connect to the simulator and use your model to test your car. To do so run drive.py file which will help you to connect with the simulator and check your saved model if it is doing good on driving the car itself.

Feel free to make play around the parameters of the code and get better and optimized results.